test: moves check functions to test_utils and adds test in transform_test.py
