+++ tests/trestle_fedramp/core/docx_helper_test.py
@@ -15,63 +15,17 @@
 # limitations under the License.
 """Testing docx utility classes."""
 
-from typing import List
-
 from docx.document import Document as DocxDocument  # type: ignore
-from docx.table import _Cell  # type: ignore
-from docx.text.paragraph import Paragraph  # type: ignore
 
 import pytest
+
+from tests.test_utils import verify_checkboxes
 
 from trestle.common.err import TrestleError
 
 from trestle_fedramp import const
 from trestle_fedramp.core.docx_helper import ControlImplementationDescriptions, ControlSummaries, FedrampDocx
 from trestle_fedramp.core.ssp_reader import FedrampControlDict, FedrampSSPData
-
-
-def verify_checkboxes(cell: _Cell, ssp_data: FedrampSSPData) -> None:
-    """Verify the checkboxes are populated correctly."""
-    checked_list: List[int] = []
-    checked_list_text: str = ''
-    for i, paragraph in enumerate(cell.paragraphs):
-        if checkbox_is_set(paragraph) and checkbox_text_is_set(paragraph):
-            checked_list.append(i)
-            checked_list_text += paragraph.text
-
-    if ssp_data.control_origination is None:
-        assert len(checked_list) == 0
-    else:
-        expected_checklist_list: List[int] = []
-        for control_origination in ssp_data.control_origination:
-            index_loc: int = ControlSummaries.get_control_origination_index(control_origination)
-            expected_checklist_list.append(index_loc)
-
-            # Check that the actual text is correct in the paragraph
-            # Each FedRAMP long string should be in the checked paragraphs of the
-            # cell.
-            assert control_origination in checked_list_text
-
-        assert checked_list == expected_checklist_list
-
-
-def checkbox_is_set(paragraph: Paragraph) -> bool:
-    """Get the checkbox value."""
-    checkboxes = paragraph._element.xpath(const.CHECKBOX_XPATH)
-    if checkboxes:
-        checkbox = checkboxes[0]
-        checked = checkbox.find(f'{const.XML_NAMESPACE}checked')
-        return checked.attrib[f'{const.XML_NAMESPACE}val'] == '1'
-    return False
-
-
-def checkbox_text_is_set(paragraph: Paragraph) -> bool:
-    """Get the checkbox text value."""
-    checkboxes = paragraph._element.xpath(const.BOX_ICON_XPATH)
-    if checkboxes:
-        checkbox = checkboxes[0]
-        return checkbox.text == const.CHECKED_BOX_ICON
-    return False
 
 
 def test_fedramp_docx_populate(docx_document: DocxDocument, test_ssp_control_dict: FedrampControlDict) -> None:
